presensi_cv/requirements.txt
opencv-contrib-python
opencv-python
numpy
pillow
mysql-connector-python
tk